requirements.txt
customtkinter
# Pillow for polaroid composition
Pillow
# PDF preview rendering
PyMuPDF
# Install tectonic as a system binary (snap or local ./bin).
